Split expectation headers
