+++ include/iris/x4/parse_result.hpp
@@ -9,8 +9,11 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
+// This header is intended for inclusion by user-facing, non-parser headers.
+// Do not add includes specific to X4's parser implementation here.
+
 #include <iris/config.hpp>
-#include <iris/x4/core/expectation.hpp>
+#include <iris/x4/core/expectation_failure.hpp>
 #include <iris/x4/core/char_traits.hpp>
 
 #include <iterator>
